refactor: rename files/tokens and expose only addDemoActionButton from factory
