+++ packages/react-components/react-storybook-addon-codesandbox/README.md
@@ -2,6 +2,6 @@
 
 **React Storybook Addon Codesandbox for [Fluent UI React](https://developer.microsoft.com/en-us/fluentui)**
 
-Storybook addon that enables codesandbox exports for stories
+Storybook addon that enables codesandbox or stackblitz exports for stories
 
 TODO: Add more details!
