+++ packages/react-components/react-storybook-addon-codesandbox/src/withCodeSandboxButton.ts
@@ -1,10 +1,9 @@
-import { /* StoryContext, */ useEffect } from '@storybook/addons';
-import { addActionButton } from './sandbox-factory';
+import { useEffect } from '@storybook/addons';
+import { addDemoActionButton } from './sandbox-factory';
 
-import { prepareData, prepareSandboxContainer } from './sandbox-utils';
 import { StoryContext } from './types';
 
-export const withCodeSandboxButton = (storyFn: (context: StoryContext) => JSX.Element, context: StoryContext) => {
+export const withSandboxButton = (storyFn: (context: StoryContext) => JSX.Element, context: StoryContext) => {
   useEffect(() => {
     if (context.viewMode === 'docs') {
       addDemoActionButton(context);
@@ -13,13 +12,3 @@ export const withCodeSandboxButton = (storyFn: (context: StoryContext) => JSX.El
 
   return storyFn(context);
 };
-
-function addDemoActionButton(context: StoryContext) {
-  const { container, cssClasses } = prepareSandboxContainer(context);
-  const config = prepareData(context);
-  if (!config) {
-    throw new Error('issues with data');
-  }
-
-  addActionButton(container, config, cssClasses);
-}
